Remove outdated testing sections for bookmark display and sorting logic
